link local addr: Do not use xor2 on the first octet of MAC
Adopts the functions that generates a link-local ipv6 address based on a EUI64 identifier to get the same result as the server-side Salt code using the `mac2eui64` function from salt.utils.network.

Whether this is ultimately correct is uncertain. We assume/hope the Salt codebase to do the right thin.